Revamp lattice surgery impl in lattice_surgery_error_detection.py
Let the file use the implementation in steane_code.py.
This change improves the error rate for initial value \ket{+},
but regresses the error rate for initial value \ket{0}.

We haven't replaced the implementation in
lattice_surgery_complementary_gap.py yet, because that regresses
the error rate.

This change also contains bug fixes and performance improvements
for the lattice surgery implementation in steane_code.py.