Introduce X_ba5 syndrome measurement
This could potentially improve the logical error rate, but in
practice this regress the logical error rate.

Note that the regression does not come from errors in the
additional syndrome extraction circuit, but from the additional
detector itself. Most of the regression goes away when we leave
the additional circuit and remove the detector.

This implies we can improve the logical error rate by modifying
the detector error model, but I don't know how...