Compress lookup tables when the syndrome condition is simple
When a gap threshold is high, it is possible that all samples with
a non-zero detector during the lattice surgery process should be
rejected. In this case, a large lookup table is unnecessary.
Instead, all samples with all-zero detectors should be accepted,
and all others rejected. This change implements the behavior.